Clean usage func generatePageFullHtml
